OJS 3.5 Compatibility - rename, move, reformat code, whitespace cleanup
