+++ assets/examples/manuscript.xml
