--- index.php
@@ -1,20 +1,0 @@
-<?php
-
-/**
- * @defgroup plugins_generic_Texture
- */
-
-/**
- * @file plugins/generic/texture/index.php
- *
- * Copyright (c) 2003-2019 Simon Fraser University
- * Copyright (c) 2003-2019 John Willinsky
- * Distributed under the GNU GPL v2. For full terms see the file docs/COPYING.
- *
- * @ingroup plugins_generic_texture
- * @brief Wrapper for Texture plugin.
- *
- */
-require_once('TexturePlugin.inc.php');
-
-return new TexturePlugin();
