+++ locale/fr/locale.po
