+++ locale/gl/locale.po
