+++ locale/hu/locale.po
