+++ locale/ka/locale.po
