+++ locale/nb/locale.po
