+++ locale/sl/locale.po
