+++ locale/sv/locale.po
